Change Config model to Configurations
Why:

While it hadn't affected operations before, when I tried to add methods
to the Config model, I received an error about RbConfig:Module. It seems
as if the Config class was formerly used in Rails, and is deprecated but
not removed. So, by trying to reference the Config model,

This change addresses the need by:

* Changing Config to Configurations everywhere: model,
migration, dashboard, routes.

* Adds a report to the seeds file, and displays an error if the
environment variables for getting a report are not yet set.

* Fixing the database config so you can drop databases. The test
database was left as SQLite, and that threw an error when running `rake
db: drop`.