Redesign email
Why:

We wanted to make the email easier to read, using MAPC's 2016 calendar
as design inspiration.

This change addresses the need by:

* Converting all templates from ERB to HAML.

* Laying out email template with a clean and readable design.

* Adding display logic to ApplicationHelper, and including it in the
mailer. Not the most appropriate place for it, but it's fine at this
scale.

* Moving recipients, sender email addresses from hardcoded values to
environment values. We don't want to need to redeploy code just to
change this value.

* Ordering a report's forecasts by created_at. (I updated a forecast
manually, and that forecast skipped to one end, because the default
ordering is updated_at.)