Content / layout changes for email
Why:

Patrick and Neal had some ideas for a layout changes. Some of the issues
stem from Outlook treating table rendering oddly, but the rest are
simple.

The requested changes were:

* Remove the top line, since it duplicates the subject.

* Make the "Today's risk" header a little smaller. I ended up rewording
so it appears as an easily understandable sentence, and is short enough
for the email preview on phones.

* Add load-shedding tips link.

* Make disclaimer text smaller

* Remove 'PEAK' from bottom row of risk levels.

* Add a background color to days when the risk is "possible". These days
are secondarily important, and we should highlight them more strongly so
people are aware of and prepared for these days.