Making style changes.
* Ensure header fonts are all sans-serif.

* Center all the table text

* Make the dates black instead of gray, for consistency

* Move the "Reference" header to below the links and above the reference
table.

* 'system2' is a typo, change to 'system'.

I tried to fix the space between the risk level on the high-risk days,
with no luck. I added a border-top and border-bottom to all those cells,
but when rendering, Outlook only applies these properties to the cell in
the first column, even though 'view source' shows the border properties
are applied to all of them.